+++ apps/meteor/ee/app/livechat-enterprise/server/lib/routing/LoadRotation.ts
@@ -30,14 +30,16 @@ class LoadRotation {
 	}
 
 	public async getNextAgent(department?: string, ignoreAgentId?: string): Promise<IOmnichannelCustomAgent | undefined> {
+		const enabledWhenIdle = settings.get<boolean>('Livechat_enabled_when_agent_idle');
+
 		const extraQuery = await getChatLimitsQuery(department);
-		const unavailableUsers = await Users.getUnavailableAgents(department, extraQuery);
-		logger.debug({ msg: 'Ignoring unavailable agents from assignment', unavailableUsers, department });
+		const unavailableUsers = await Users.getUnavailableAgents(department, extraQuery, enabledWhenIdle);
+		logger.debug({ msg: 'Ignoring unavailable agents from assignment', unavailableUsers, department, enabledWhenIdle });
 
 		const nextAgent = await Users.getLastAvailableAgentRouted(
 			department,
 			ignoreAgentId,
-			settings.get<boolean>('Livechat_enabled_when_agent_idle'),
+			enabledWhenIdle,
 			unavailableUsers.map((user) => user.username),
 		);
 		if (!nextAgent?.username) {