+++ packages/model-typings/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/model-typings
 
+## 2.1.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.2.1
+  </details>
+
 ## 2.1.0
 
 ### Minor Changes